+++ custom_inputs/bme280 serial to ttn/mycodo_custom_input_bme280_ttn.py
@@ -53,9 +53,9 @@ measurements_dict = {
 
 # Input information
 INPUT_INFORMATION = {
-    'input_name_unique': 'BME280_TTN',
+    'input_name_unique': 'BME280_TTN_1_1',
     'input_manufacturer': 'BOSCH',
-    'input_name': 'BME280 (->Serial->TTN)',
+    'input_name': 'BME280 (->Serial->TTN) v1.1',
     'input_library': 'Adafruit_BME280/pyserial',
     'measurements_name': 'Pressure/Humidity/Temperature',
     'measurements_dict': measurements_dict,
@@ -174,8 +174,7 @@ class InputModule(AbstractInput):
                 # "B" designates this data belonging to the BME280
                 string_send = 'B,{},{},{}'.format(self.value_get(1), self.value_get(2), self.value_get(0))
                 if self.serial_device:
-                    self.lock_acquire(self.lock_file, timeout=10)
-                    if self.locked[self.lock_file]:
+                    if self.lock_acquire(self.lock_file, timeout=10):
                         try:
                             self.serial_send = self.serial.Serial(self.serial_device, 9600)
                             self.serial_send.write(string_send.encode())