+++ custom_outputs/remote GPIO on-off/mycodo_custom_output_remote_gpio_on_off.py
@@ -13,13 +13,20 @@ measurements_dict = {
     }
 }
 
+channels_dict = {
+    0: {
+        'types': ['on_off'],
+        'measurements': [0]
+    }
+}
+
 # Output information
 OUTPUT_INFORMATION = {
-    'output_name_unique': 'remote_gpio_on_off',
-    'output_name': lazy_gettext('On/Off Remote GPIO'),
+    'output_name_unique': 'remote_gpio_on_off_1_1',
+    'output_name': "{} v1.1".format(lazy_gettext('On/Off Remote GPIO')),
     'output_library': 'gpiozero',
     'measurements_dict': measurements_dict,
-    'version': '0.1',
+    'channels_dict': channels_dict,
 
     'on_state_internally_handled': False,