+++ custom_inputs/bme280 serial to ttn/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1 (2021-02-08)
+
+ - Update locking to work with latest Mycodo
+
 ## 1.0 (2020-06-17)
 
 Original Release