+++ custom_outputs/remote GPIO on-off/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.1 (2020-05-14)
+## 1.1 (2021-02-08)
+
+ - Update channels to work with latest Mycodo
+
+## 1.0 (2020-05-14)
 
 Original Release (currently untested)